backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.3.0
dj-database-url==2.1.0
pytest
pytest-django==4.7.0
coverage==7.3.2